Fix M117 (Display message) throwing an exception and added the IDictionary interface to CommandBase
The M117 command does not follow the same format as most of the other
commands. It allows whitespace, has a string instead of numbers, ends at
the new line, and does not contain a parameter type before the value.

Also, added the IDictionary interface to CommandBase because CommandBase
is a set of Key/Value pairs.